+++ lib/src/cached_image_widget.dart
@@ -1,6 +1,4 @@
 import 'dart:async';
-import 'dart:io';
-import 'dart:ui' as ui;
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
@@ -11,8 +9,10 @@ import 'package:flutter_cache_manager/flutter_cache_manager.dart';
 typedef Widget ImageWidgetBuilder(
     BuildContext context, ImageProvider imageProvider);
 typedef Widget PlaceholderWidgetBuilder(BuildContext context, String url);
+typedef Widget ProgressIndicatorBuilder(
+    BuildContext context, String url, DownloadProgress progress);
 typedef Widget LoadingErrorWidgetBuilder(
-    BuildContext context, String url, Object error);
+    BuildContext context, String url, dynamic error);
 
 class CachedNetworkImage extends StatefulWidget {
   /// Option to use cachemanager with other settings
@@ -26,6 +26,9 @@ class CachedNetworkImage extends StatefulWidget {
 
   /// Widget displayed while the target [imageUrl] is loading.
   final PlaceholderWidgetBuilder placeholder;
+
+  /// Widget displayed while the target [imageUrl] is loading.
+  final ProgressIndicatorBuilder progressIndicatorBuilder;
 
   /// Widget displayed while the target [imageUrl] failed loading.
   final LoadingErrorWidgetBuilder errorWidget;
@@ -131,49 +134,55 @@ class CachedNetworkImage extends StatefulWidget {
   ///  * [BlendMode], which includes an illustration of the effect of each blend mode.
   final BlendMode colorBlendMode;
 
-  final bool playGif;
+  /// Target the interpolation quality for image scaling.
+  ///
+  /// If not given a value, defaults to FilterQuality.low.
+  final FilterQuality filterQuality;
 
   CachedNetworkImage({
     Key key,
     @required this.imageUrl,
     this.imageBuilder,
     this.placeholder,
+    this.progressIndicatorBuilder,
     this.errorWidget,
-    this.fadeOutDuration: const Duration(milliseconds: 1000),
-    this.fadeOutCurve: Curves.easeOut,
-    this.fadeInDuration: const Duration(milliseconds: 500),
-    this.fadeInCurve: Curves.easeIn,
+    this.fadeOutDuration = const Duration(milliseconds: 1000),
+    this.fadeOutCurve = Curves.easeOut,
+    this.fadeInDuration = const Duration(milliseconds: 500),
+    this.fadeInCurve = Curves.easeIn,
     this.width,
     this.height,
     this.fit,
-    this.alignment: Alignment.center,
-    this.repeat: ImageRepeat.noRepeat,
-    this.matchTextDirection: false,
+    this.alignment = Alignment.center,
+    this.repeat = ImageRepeat.noRepeat,
+    this.matchTextDirection = false,
     this.httpHeaders,
     this.cacheManager,
-    this.useOldImageOnUrlChange: false,
+    this.useOldImageOnUrlChange = false,
     this.color,
+    this.filterQuality = FilterQuality.low,
     this.colorBlendMode,
     this.placeholderFadeInDuration,
-    this.playGif: true,
   })  : assert(imageUrl != null),
         assert(fadeOutDuration != null),
         assert(fadeOutCurve != null),
         assert(fadeInDuration != null),
         assert(fadeInCurve != null),
         assert(alignment != null),
+        assert(filterQuality != null),
         assert(repeat != null),
         assert(matchTextDirection != null),
         super(key: key);
 
   @override
   CachedNetworkImageState createState() {
-    return new CachedNetworkImageState();
+    return CachedNetworkImageState();
   }
 }
 
 class _ImageTransitionHolder {
   final FileInfo image;
+  final DownloadProgress progress;
   AnimationController animationController;
   final Object error;
   Curve curve;
@@ -181,12 +190,13 @@ class _ImageTransitionHolder {
 
   _ImageTransitionHolder({
     this.image,
+    this.progress,
     @required this.animationController,
     this.error,
-    this.curve: Curves.easeIn,
+    this.curve = Curves.easeIn,
   }) : forwardTickerFuture = animationController.forward();
 
-  dispose() {
+  void dispose() {
     if (animationController != null) {
       animationController.dispose();
       animationController = null;
@@ -196,8 +206,10 @@ class _ImageTransitionHolder {
 
 class CachedNetworkImageState extends State<CachedNetworkImage>
     with TickerProviderStateMixin {
-  List<_ImageTransitionHolder> _imageHolders = List();
+  final _imageHolders = <_ImageTransitionHolder>[];
   Key _streamBuilderKey = UniqueKey();
+  Stream<FileResponse> _fileResponseStream;
+  FileInfo _fromMemory;
 
   @override
   Widget build(BuildContext context) {
@@ -205,12 +217,20 @@ class CachedNetworkImageState extends State<CachedNetworkImage>
   }
 
   @override
+  void initState() {
+    super.initState();
+    _createFileStream();
+  }
+
+  @override
   void didUpdateWidget(CachedNetworkImage oldWidget) {
-    if (oldWidget.imageUrl != widget.imageUrl &&
-        !widget.useOldImageOnUrlChange) {
+    if (oldWidget.imageUrl != widget.imageUrl) {
       _streamBuilderKey = UniqueKey();
-      _disposeImageHolders();
-      _imageHolders.clear();
+      if (!widget.useOldImageOnUrlChange) {
+        _disposeImageHolders();
+        _imageHolders.clear();
+      }
+      _createFileStream();
     }
     super.didUpdateWidget(oldWidget);
   }
@@ -221,44 +241,74 @@ class CachedNetworkImageState extends State<CachedNetworkImage>
     super.dispose();
   }
 
-  _disposeImageHolders() {
+  void _createFileStream() {
+    _fromMemory = _cacheManager().getFileFromMemory(widget.imageUrl);
+
+    _fileResponseStream = _cacheManager()
+        .getFileStream(
+          widget.imageUrl,
+          headers: widget.httpHeaders,
+          withProgress: widget.progressIndicatorBuilder != null,
+        )
+        // ignore errors if not mounted
+        .handleError(() {}, test: (_) => !mounted)
+        .where((f) {
+      if (f is FileInfo) {
+        return f?.originalUrl != _fromMemory?.originalUrl ||
+            f?.validTill != _fromMemory?.validTill;
+      }
+      return true;
+    });
+  }
+
+  void _disposeImageHolders() {
     for (var imageHolder in _imageHolders) {
       imageHolder.dispose();
     }
   }
 
-  _addImage({FileInfo image, Object error, Duration duration}) {
-    if (_imageHolders.length > 0) {
+  void _addImage(
+      {FileInfo image,
+      DownloadProgress progress,
+      Object error,
+      Duration duration}) {
+    if (_imageHolders.isNotEmpty) {
       var lastHolder = _imageHolders.last;
-      lastHolder.forwardTickerFuture.then((_) {
-        if (lastHolder.animationController == null) {
-          return;
-        }
-        if (widget.fadeOutDuration != null) {
-          lastHolder.animationController.duration = widget.fadeOutDuration;
-        } else {
-          lastHolder.animationController.duration = Duration(seconds: 1);
-        }
-        if (widget.fadeOutCurve != null) {
-          lastHolder.curve = widget.fadeOutCurve;
-        } else {
-          lastHolder.curve = Curves.easeOut;
-        }
-        lastHolder.animationController.reverse().then((_) {
-          _imageHolders.remove(lastHolder);
-          if (mounted) setState(() {});
-          return null;
+      if (lastHolder.progress != null && progress != null) {
+        _imageHolders.removeLast();
+      } else {
+        lastHolder.forwardTickerFuture.then((_) {
+          if (lastHolder.animationController == null) {
+            return;
+          }
+          if (widget.fadeOutDuration != null) {
+            lastHolder.animationController.duration = widget.fadeOutDuration;
+          } else {
+            lastHolder.animationController.duration =
+                const Duration(seconds: 1);
+          }
+          if (widget.fadeOutCurve != null) {
+            lastHolder.curve = widget.fadeOutCurve;
+          } else {
+            lastHolder.curve = Curves.easeOut;
+          }
+          lastHolder.animationController.reverse().then((_) {
+            _imageHolders.remove(lastHolder);
+            if (mounted) setState(() {});
+            return null;
+          });
         });
-      });
+      }
     }
     _imageHolders.add(
       _ImageTransitionHolder(
         image: image,
         error: error,
+        progress: progress,
         animationController: AnimationController(
           vsync: this,
           duration: duration ??
-              (widget.fadeInDuration ?? Duration(milliseconds: 500)),
+              (widget.fadeInDuration ?? const Duration(milliseconds: 500)),
         ),
       ),
     );
@@ -268,38 +318,46 @@ class CachedNetworkImageState extends State<CachedNetworkImage>
     return _cacheManager().getFileFromMemory(widget.imageUrl);
   }
 
-  _animatedWidget() {
-    var fromMemory = _getFromMemory();
-
-    return StreamBuilder<FileInfo>(
+  Widget _animatedWidget() {
+    return StreamBuilder<FileResponse>(
       key: _streamBuilderKey,
-      initialData: fromMemory,
-      stream: _cacheManager()
-          .getFile(widget.imageUrl, headers: widget.httpHeaders)
-          .where((f) =>
-              f?.originalUrl != fromMemory?.originalUrl ||
-              f?.validTill != fromMemory?.validTill),
-      builder: (BuildContext context, AsyncSnapshot<FileInfo> snapshot) {
+      initialData: _fromMemory,
+      stream: _fileResponseStream,
+      builder: (BuildContext context, AsyncSnapshot<FileResponse> snapshot) {
         if (snapshot.hasError) {
           // error
-          if (_imageHolders.length == 0 || _imageHolders.last.error == null) {
+          if (_imageHolders.isEmpty || _imageHolders.last.error == null) {
             _addImage(image: null, error: snapshot.error);
           }
         } else {
-          var fileInfo = snapshot.data;
-          if (fileInfo == null) {
+          var fileResponse = snapshot.data;
+          if (fileResponse == null) {
             // placeholder
-            if (_imageHolders.length == 0 || _imageHolders.last.image != null) {
+            if (_imageHolders.isEmpty || _imageHolders.last.image != null) {
+              DownloadProgress progress;
+              if (widget.progressIndicatorBuilder != null) {
+                progress = DownloadProgress(widget.imageUrl, null, 0);
+              }
               _addImage(
+                  progress: progress,
                   image: null,
                   duration: widget.placeholderFadeInDuration ?? Duration.zero);
             }
-          } else if (_imageHolders.length == 0 ||
-              _imageHolders.last.image?.originalUrl != fileInfo.originalUrl ||
-              _imageHolders.last.image?.validTill != fileInfo.validTill) {
-            _addImage(
-                image: fileInfo,
-                duration: _imageHolders.length > 0 ? null : Duration.zero);
+          } else {
+            if (fileResponse is FileInfo) {
+              if (_imageHolders.isEmpty ||
+                  _imageHolders.last.image?.originalUrl !=
+                      fileResponse.originalUrl ||
+                  _imageHolders.last.image?.validTill !=
+                      fileResponse.validTill) {
+                _addImage(
+                    image: fileResponse,
+                    duration: _imageHolders.isNotEmpty ? null : Duration.zero);
+              }
+            }
+            if (fileResponse is DownloadProgress) {
+              _addImage(progress: fileResponse, duration: Duration.zero);
+            }
           }
         }
 
@@ -308,15 +366,26 @@ class CachedNetworkImageState extends State<CachedNetworkImage>
           if (holder.error != null) {
             children.add(_transitionWidget(
                 holder: holder, child: _errorWidget(context, holder.error)));
+          } else if (holder.progress != null) {
+            children.add(_transitionWidget(
+                holder: holder,
+                child: widget.progressIndicatorBuilder(
+                  context,
+                  holder.progress.originalUrl,
+                  holder.progress,
+                )));
           } else if (holder.image == null) {
             children.add(_transitionWidget(
                 holder: holder, child: _placeholder(context)));
           } else {
             children.add(_transitionWidget(
                 holder: holder,
-                child: _image(
-                  context,
-                  FileImage(holder.image.file),
+                child: KeyedSubtree(
+                  key: Key(holder.image.file.path),
+                  child: _image(
+                    context,
+                    FileImage(holder.image.file),
+                  ),
                 )));
           }
         }
@@ -338,55 +407,37 @@ class CachedNetworkImageState extends State<CachedNetworkImage>
     );
   }
 
-  _cacheManager() {
+  BaseCacheManager _cacheManager() {
     return widget.cacheManager ?? DefaultCacheManager();
   }
 
-  _image(BuildContext context, ImageProvider imageProvider) {
-    if (widget.playGif) {
-      return widget.imageBuilder != null
-          ? widget.imageBuilder(context, imageProvider)
-          : new Image(
-              image: imageProvider,
-              fit: widget.fit,
-              width: widget.width,
-              height: widget.height,
-              alignment: widget.alignment,
-              repeat: widget.repeat,
-              color: widget.color,
-              colorBlendMode: widget.colorBlendMode,
-              matchTextDirection: widget.matchTextDirection,
-            );
-    } else {
-      return FutureBuilder(
-        builder: (_, _snapshot) => RawImage(
-          image: _snapshot.data,
-        ),
-        future: _getGifFirstFrame(imageProvider),
-      );
-    }
+  Widget _image(BuildContext context, ImageProvider imageProvider) {
+    return widget.imageBuilder != null
+        ? widget.imageBuilder(context, imageProvider)
+        : Image(
+            image: imageProvider,
+            fit: widget.fit,
+            width: widget.width,
+            height: widget.height,
+            alignment: widget.alignment,
+            repeat: widget.repeat,
+            color: widget.color,
+            colorBlendMode: widget.colorBlendMode,
+            matchTextDirection: widget.matchTextDirection,
+            filterQuality: widget.filterQuality,
+          );
   }
 
-  Future<ui.Image> _getGifFirstFrame(ImageProvider provider) async {
-    File file = await _cacheManager().getSingleFile(widget.imageUrl);
-    var data = file.readAsBytesSync();
-
-    ui.Codec codec = await PaintingBinding.instance
-        .instantiateImageCodec(data.buffer.asUint8List());
-    ui.FrameInfo frameInfo = await codec.getNextFrame();
-    return Future<ui.Image>.value(frameInfo.image);
-  }
-
-  _placeholder(BuildContext context) {
+  Widget _placeholder(BuildContext context) {
     return widget.placeholder != null
         ? widget.placeholder(context, widget.imageUrl)
-        : new SizedBox(
+        : SizedBox(
             width: widget.width,
             height: widget.height,
           );
   }
 
-  _errorWidget(BuildContext context, Object error) {
+  Widget _errorWidget(BuildContext context, Object error) {
     return widget.errorWidget != null
         ? widget.errorWidget(context, widget.imageUrl, error)
         : _placeholder(context);