+++ lib/src/cached_network_image_provider.dart
@@ -1,6 +1,5 @@
 import 'dart:async' show Future;
 import 'dart:io' show File;
-import 'dart:typed_data';
 import 'dart:ui' as ui show instantiateImageCodec, Codec;
 
 import 'package:flutter/foundation.dart';
@@ -10,18 +9,18 @@ import 'package:flutter_cache_manager/flutter_cache_manager.dart';
 typedef void ErrorListener();
 typedef CompressCallback = Future<File> Function(File);
 
-class CachedNetworkImageProvider
-    extends ImageProvider<CachedNetworkImageProvider> {
+class CachedNetworkImageProvider extends ImageProvider<CachedNetworkImageProvider> {
   /// Creates an ImageProvider which loads an image from the [url], using the [scale].
   /// When the image fails to load [errorListener] is called.
-  const CachedNetworkImageProvider(this.url,
-      {this.scale: 1.0,
-      this.errorListener,
-      this.headers,
-      this.cacheManager,
-      this.compressCallback,
-      this.isDeleteSourceCached: false})
-      : assert(url != null),
+  const CachedNetworkImageProvider(
+    this.url, {
+    this.scale: 1.0,
+    this.errorListener,
+    this.headers,
+    this.cacheManager,
+    this.compressCallback,
+    this.isDeleteSourceCached = true,
+  })  : assert(url != null),
         assert(scale != null);
 
   final BaseCacheManager cacheManager;
@@ -43,25 +42,22 @@ class CachedNetworkImageProvider
   final bool isDeleteSourceCached;
 
   @override
-  Future<CachedNetworkImageProvider> obtainKey(
-      ImageConfiguration configuration) {
-    return new SynchronousFuture<CachedNetworkImageProvider>(this);
+  Future<CachedNetworkImageProvider> obtainKey(ImageConfiguration configuration) {
+    return SynchronousFuture<CachedNetworkImageProvider>(this);
   }
 
   @override
-  ImageStreamCompleter load(
-      CachedNetworkImageProvider key, DecoderCallback decode) {
-    return new MultiFrameImageStreamCompleter(
+  ImageStreamCompleter load(CachedNetworkImageProvider key, DecoderCallback decode) {
+    return MultiFrameImageStreamCompleter(
       codec: _loadAsync(key),
       scale: key.scale,
-// TODO enable information collector on next stable release of flutter
-//      informationCollector: () sync* {
-//        yield DiagnosticsProperty<ImageProvider>(
-//          'Image provider: $this \n Image key: $key',
-//          this,
-//          style: DiagnosticsTreeStyle.errorProperty,
-//        );
-//      },
+      informationCollector: () sync* {
+        yield DiagnosticsProperty<ImageProvider>(
+          'Image provider: $this \n Image key: $key',
+          this,
+          style: DiagnosticsTreeStyle.errorProperty,
+        );
+      },
     );
   }
 
@@ -70,29 +66,29 @@ class CachedNetworkImageProvider
     var file = await mngr.getSingleFile(url, headers: headers);
     if (file == null) {
       if (errorListener != null) errorListener();
-      return Future<ui.Codec>.error("Couldn't download or retrieve file.");
+      throw Exception('Couldn\'t download or retrieve file: $url');
     }
     if (compressCallback != null) {
       file = await _compress(file, mngr);
       if (file == null) {
         if (errorListener != null) errorListener();
-        return Future<ui.Codec>.error("Image compression failed.");
+        throw Exception('Image compression failed.');
       }
     }
-    return await _loadAsyncFromFile(key, file);
+    return _loadAsyncFromFile(key, file);
   }
 
-  Future<ui.Codec> _loadAsyncFromFile(
-      CachedNetworkImageProvider key, File file) async {
+  Future<ui.Codec> _loadAsyncFromFile(CachedNetworkImageProvider key, File file) async {
     assert(key == this);
 
-    final Uint8List bytes = await file.readAsBytes();
+    final bytes = await file.readAsBytes();
 
     if (bytes.lengthInBytes == 0) {
       if (errorListener != null) errorListener();
-      throw new Exception("File was empty");
+      throw Exception('File was empty');
     }
-    return await ui.instantiateImageCodec(bytes);
+
+    return ui.instantiateImageCodec(bytes);
   }
 
   Future<File> _compress(File file, BaseCacheManager mngr) async {
@@ -101,7 +97,7 @@ class CachedNetworkImageProvider
     var result;
     if (fileInfo == null) {
       result = await compressCallback(file);
-      mngr.putFile(tempUrl, result.readAsBytesSync());
+      await mngr.putFile(tempUrl, result.readAsBytesSync());
       if (isDeleteSourceCached) {
         await mngr.removeFile(url);
       }
@@ -113,9 +109,10 @@ class CachedNetworkImageProvider
 
   @override
   bool operator ==(dynamic other) {
-    if (other.runtimeType != runtimeType) return false;
-    final CachedNetworkImageProvider typedOther = other;
-    return url == typedOther.url && scale == typedOther.scale;
+    if (other is CachedNetworkImageProvider) {
+      return url == other.url && scale == other.scale;
+    }
+    return false;
   }
 
   @override